+++ feincms3/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (0, 90, 0)
+VERSION = (0, 91, 0)
 __version__ = ".".join(map(str, VERSION))