+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+
+`0.91`_ (2021-05-28)
+~~~~~~~~~~~~~~~~~~~~
+
 Inline CKEditor widget
 ----------------------
 
@@ -804,4 +808,5 @@ functionality.
 .. _0.40: https://github.com/matthiask/feincms3/compare/0.39...0.40
 .. _0.41: https://github.com/matthiask/feincms3/compare/0.40...0.41
 .. _0.90: https://github.com/matthiask/feincms3/compare/0.41...0.90
-.. _Next version: https://github.com/matthiask/feincms3/compare/0.90...main
+.. _0.91: https://github.com/matthiask/feincms3/compare/0.90...0.91
+.. _Next version: https://github.com/matthiask/feincms3/compare/0.91...main